Fixing various type warnings. We don't today have these warnings enabled in our code but it'd be nice to remove them from the list at some point. This starts us down that path by hopefully making it easier for code authors to copy/paste or at least notice the right patterns.

+++ iree/hal/vulkan/descriptor_pool_cache.cc
@@ -65,7 +65,7 @@ StatusOr<DescriptorPool> DescriptorPoolCache::AcquireDescriptorPool(
   std::array<VkDescriptorPoolSize, 1> pool_sizes;
   pool_sizes[0].type = descriptor_type;
   pool_sizes[0].descriptorCount = max_descriptor_count;
-  create_info.poolSizeCount = pool_sizes.size();
+  create_info.poolSizeCount = static_cast<uint32_t>(pool_sizes.size());
   create_info.pPoolSizes = pool_sizes.data();
 
   DescriptorPool descriptor_pool;
